requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
sqlalchemy>=2.0.0
yfinance>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit-aggrid>=0.3.4
openpyxl>=3.1.0